+++ CHANGELOG.md
@@ -6,13 +6,15 @@ For the detailed description, please explore nested folders and corresponding CH
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.3] - 2026-08-11
+
+## [1.2.4] - 2026-08-31
 
 ### Added
 
 ### Fixed
 
 - Acquisition PCB layout changes to improve RF performance
+- Firmware improvements on nRF52 to improved reliability
 
 ### Changed